Extract UI into subcomponents
